refact some functions declaration and interfaces
